+++ src/main/java/com/example/college/enums/AcademicTitle.java
@@ -1,6 +1,6 @@
 package com.example.college.enums;
 
 public enum AcademicTitle {
-    DOCTOR,
+    PHD,
     PROFESSOR
 }
