+++ src/main/java/com/example/college/major/MajorService.java
@@ -5,25 +5,14 @@ import com.example.college.subject.SubjectRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
-import static com.example.college.randomhelper.RandomHelper.generateRandomNumberFromRange;
-
 @Service
 @RequiredArgsConstructor
 public class MajorService {
     private final MajorRepository majorRepository;
     private final SubjectRepository subjectRepository;
-    public void addSubjectsToAllMajors() {
-        majorRepository.findAll().forEach(major -> {
-                addSubject(major, "Przyrka");
-                addSubject(major, "Matma");
-                addSubject(major, "WOS");
-        });
-    }
 
-    private void addSubject(Major major, String name) {
-        Subject subject = new Subject(name,generateRandomNumberFromRange(1,10));
+    public void createSubjectAddToMajor(Major major, String name, int amountECTS) {
+        Subject subject = new Subject(name, amountECTS);
         subject.setMajor(major);
         major.getSubjectList().add(subject);
         subjectRepository.save(subject);
